Dedupe Entity/Ego type selector
